fixed code, ready to do pull request
