+++ front-end/src/App.js
@@ -1,14 +1,13 @@
-import logo from './logo.svg';
 import './App.css';
 
 import {BrowserRouter as Router, Route, Switch, Link, Redirect} from 'react-router-dom'
 // pages
-import HomePage from "./pages/home"
+import HomePage from "./home"
 
 function App() {
   return (
     <Router>
-      <Route path='/' component={HomePage}></Route>
+      <Route exact path='/' component={HomePage}></Route>
     </Router>
   );
 }
